test(auth): neutral naming for the limiter-key pins
Renames the pins and their file to describe the property they assert rather than the failure mode, and rewords the failure output in the same terms. No change to what is exercised or asserted.
